Make controller with routes to fetch the data
